Add files via upload
This is a basic java chat app that captures user registration, stored this input from the user and allows them to eventually login using the same registration input captured and stored when user initially entered input. User has rules to follow in order for the input to be accepted by the java program.